remove old signal processor and hook in new version
